serialize: Change GetSerializeSize return type to uint64_t
GetSerializeSize currently returns size_t, which is platform-dependent.
On 32-bit systems, size_t is 32-bit, limiting the maximum serialized
size to ~4GB. This can cause integer overflow when multiplying by
WITNESS_SCALE_FACTOR during block size validation.

Change GetSerializeSize to return uint64_t to ensure it can handle
large sizes without overflow on all platforms.

This fixes CVE-2025-46597.

+++ src/coins.cpp
@@ -349,7 +349,7 @@ void CCoinsViewCache::SanityCheck() const
     assert(recomputed_usage == cachedCoinsUsage);
 }
 
-static const size_t MIN_TRANSACTION_OUTPUT_WEIGHT = WITNESS_SCALE_FACTOR * ::GetSerializeSize(CTxOut());
+static const uint64_t MIN_TRANSACTION_OUTPUT_WEIGHT = WITNESS_SCALE_FACTOR * ::GetSerializeSize(CTxOut());
 static const size_t MAX_OUTPUTS_PER_BLOCK = MAX_BLOCK_WEIGHT / MIN_TRANSACTION_OUTPUT_WEIGHT;
 
 const Coin& AccessByTxid(const CCoinsViewCache& view, const Txid& txid)
